# Sistema_Gestión_Donante/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Sistema de Gestión de Donantes - API",
        "description": "API para el módulo de donantes del sistema de adopciones de animales",
        "contact": {
            "email": "admin@sistema-adopciones.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor de desarrollo"
        }
    ],
    "paths": {
...
        "/api/historial-colaboraciones": {
            "get": {
                "tags": [
                    "Historial Colaboraciones"
                ],
                "summary": "Obtener todo el historial de colaboraciones",
                "operationId": "025b856866568a52854855ebe2bb669c",
                "responses": {
                    "200": {
                        "description": "Lista de historial de colaboraciones obtenida exitosamente"
                    },
                    "500": {
                        "description": "Error del servidor"
                    }
                }
            },
            "post": {
                "tags": [
                    "Historial Colaboraciones"
                ],
                "summary": "Crear una nueva colaboración",
                "operationId": "a45371c924c1edd941e1637ce888ac5b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "donante_id",
                                    "tipo_colaboracion",
                                    "descripcion",
                                    "fecha_colaboracion"
                                ],
                                "properties": {
                                    "donante_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "tipo_colaboracion": {
                                        "type": "string",
                                        "example": "Económica"
                                    },
                                    "descripcion": {
                                        "type": "string",
                                        "example": "Donación monetaria para alimento"
                                    },
                                    "fecha_colaboracion": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2023-06-08"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Colaboración creada exitosamente"
                    },
                    "400": {
                        "description": "Datos inválidos"
                    },
                    "500": {
                        "description": "Error del servidor"
                    }
                }
            }
        },
        "/api/historial-colaboraciones/{id}": {
            "get": {
                "tags": [
                    "Historial Colaboraciones"
                ],
                "summary": "Obtener una colaboración por ID",
                "operationId": "fe13728571b0a803932ae281a3f8fe34",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la colaboración",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Colaboración obtenida exitosamente"
                    },
                    "404": {
                        "description": "Colaboración no encontrada"
                    },
                    "500": {
                        "description": "Error del servidor"
                    }
                }
            },
            "put": {
                "tags": [
                    "Historial Colaboraciones"
                ],
                "summary": "Actualizar una colaboración existente",
                "operationId": "1bf952a2d0a261b4b3f5ded1cf953ad5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la colaboración",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "donante_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "tipo_colaboracion": {
                                        "type": "string",
                                        "example": "Voluntariado"
                                    },
                                    "descripcion": {
                                        "type": "string",
                                        "example": "Descripción actualizada"
                                    },
                                    "fecha_colaboracion": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2023-06-08"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Colaboración actualizada exitosamente"
                    },
                    "400": {
                        "description": "Datos inválidos"
                    },
                    "404": {
                        "description": "Colaboración no encontrada"
                    },
                    "500": {
                        "description": "Error del servidor"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Historial Colaboraciones"
                ],
                "summary": "Eliminar una colaboración",
                "operationId": "813734f37d6d2e58492e32861372fb36",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la colaboración",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Colaboración eliminada exitosamente"
                    },
                    "404": {
                        "description": "Colaboración no encontrada"
                    },
                    "500": {
                        "description": "Error del servidor"
                    }
                }
            }
        },
        "/api/historial-colaboraciones/donante/{donanteId}": {
            "get": {
                "tags": [
                    "Historial Colaboraciones"
                ],
                "summary": "Obtener colaboraciones por donante",
                "operationId": "26b6d191f49c085ecefab69a24f57f0e",
                "parameters": [
                    {
                        "name": "donanteId",
                        "in": "path",
                        "description": "ID del donante",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de colaboraciones por donante obtenida exitosamente"
                    },
                    "500": {
                        "description": "Error del servidor"
                    }
                }
            }
        },
        "/api/historial-colaboraciones/tipo/{tipo}": {
            "get": {
                "tags": [
                    "Historial Colaboraciones"
                ],
                "summary": "Obtener colaboraciones por tipo",
                "operationId": "11fe0a996633439d1869fd2cf833e71c",
                "parameters": [
                    {
                        "name": "tipo",
                        "in": "path",
                        "description": "Tipo de colaboración",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de colaboraciones por tipo obtenida exitosamente"
                    },
                    "500": {
                        "description": "Error del servidor"
                    }
                }
            }
        },
...
    },
...
}